functions/requirements.txt
firebase-functions
firebase-admin
google-generativeai
httpx
stripe
# For AADE if needed, though we used standard xml
